# perf(sql): speed up simple LIKE/ILIKE filters