--- a/benchmarks/src/main/java/org/questdb/CharsIndexOfBenchmark.java
+++ b/benchmarks/src/main/java/org/questdb/CharsIndexOfBenchmark.java
@@ -1,0 +1,83 @@
+/*******************************************************************************
+ *     ___                  _   ____  ____
+ *    / _ \ _   _  ___  ___| |_|  _ \| __ )
+ *   | | | | | | |/ _ \/ __| __| | | |  _ \
+ *   | |_| | |_| |  __/\__ \ |_| |_| | |_) |
+ *    \__\_\\__,_|\___||___/\__|____/|____/
+ *
+ *  Copyright (c) 2014-2019 Appsicle
+ *  Copyright (c) 2019-2023 QuestDB
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ ******************************************************************************/
+
+package org.questdb;
+
+import io.questdb.std.Chars;
+import io.questdb.std.NanosecondClockImpl;
+import io.questdb.std.Rnd;
+import io.questdb.std.datetime.microtime.MicrosecondClockImpl;
+import org.openjdk.jmh.annotations.*;
+import org.openjdk.jmh.runner.Runner;
+import org.openjdk.jmh.runner.RunnerException;
+import org.openjdk.jmh.runner.options.Options;
+import org.openjdk.jmh.runner.options.OptionsBuilder;
+
+import java.util.concurrent.TimeUnit;
+
+@State(Scope.Benchmark)
+@BenchmarkMode(Mode.AverageTime)
+@OutputTimeUnit(TimeUnit.NANOSECONDS)
+public class CharsIndexOfBenchmark {
+
+    @Param({"32", "128", "1024"})
+    private int len;
+    private String seq;
+
+    public static void main(String[] args) throws RunnerException {
+        Options opt = new OptionsBuilder()
+                .include(CharsIndexOfBenchmark.class.getSimpleName())
+                .warmupIterations(3)
+                .measurementIterations(3)
+                .forks(1)
+                .build();
+        new Runner(opt).run();
+    }
+
+    @Setup(Level.Iteration)
+    public void setUp() {
+        Rnd rnd = new Rnd(NanosecondClockImpl.INSTANCE.getTicks(), MicrosecondClockImpl.INSTANCE.getTicks());
+        StringBuilder sb = new StringBuilder();
+        for (int i = 0; i < len; i++) {
+            sb.append(rnd.nextChar());
+        }
+        sb.append("google yahoo facebook");
+        seq = sb.toString();
+    }
+
+    @Benchmark
+    public int testBidirectionalIndexOf() {
+        return Chars.indexOf(seq, 0, seq.length(), "google");
+    }
+
+    @Benchmark
+    public int testNaiveIndexOf() {
+        return Chars.indexOf(seq, 0, seq.length(), "google", 1);
+    }
+
+    @Benchmark
+    public int testStdlibIndexOf() {
+        return seq.indexOf("google");
+    }
+}