--- a/core/src/main/java/io/questdb/griffin/engine/functions/regex/AbstractLikeStrFunctionFactory.java
+++ b/core/src/main/java/io/questdb/griffin/engine/functions/regex/AbstractLikeStrFunctionFactory.java
@@ -41,6 +41,7 @@ import io.questdb.std.IntList;
 import io.questdb.std.Misc;
 import io.questdb.std.ObjList;
 import io.questdb.std.str.StringSink;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -53,17 +54,18 @@ public abstract class AbstractLikeStrFunctionFactory implements FunctionFactory 
         StringSink sink = Misc.getThreadLocalSink();
         for (int i = 0; i < len; i++) {
             char c = pattern.charAt(i);
-            if (c == '_')
+            if (c == '_') {
                 sink.put('.');
-            else if (c == '%')
+            } else if (c == '%') {
                 sink.put(".*?");
-            else if ("[](){}.*+?$^|#".indexOf(c) != -1) {
+            } else if ("[](){}.*+?$^|#".indexOf(c) != -1) {
                 sink.put("\\");
                 sink.put(c);
             } else if (c == '\\') {
-                i += 1;
-                if (i >= len)
+                i++;
+                if (i >= len) {
                     throw SqlException.parserErr(i - 1, pattern, "LIKE pattern must not end with escape character");
+                }
                 c = pattern.charAt(i);
                 if ("[](){}.*+?$^|#\\".indexOf(c) != -1) {
                     sink.put("\\");
@@ -71,8 +73,9 @@ public abstract class AbstractLikeStrFunctionFactory implements FunctionFactory 
                 } else {
                     sink.put(c);
                 }
-            } else
+            } else {
                 sink.put(c);
+            }
         }
 
         if (Chars.equalsNc(sink, prev)) {
@@ -93,9 +96,49 @@ public abstract class AbstractLikeStrFunctionFactory implements FunctionFactory 
         final Function pattern = args.getQuick(1);
 
         if (pattern.isConstant()) {
-            final CharSequence likeString = pattern.getStr(null);
-            if (likeString != null && likeString.length() > 0) {
-                String p = escapeSpecialChars(likeString, null);
+            final CharSequence likeSeq = pattern.getStr(null);
+            int len;
+            if (likeSeq != null && (len = likeSeq.length()) > 0) {
+                int oneCount = countChar(likeSeq, '_');
+                if (oneCount == 0) {
+                    if (likeSeq.charAt(len - 1) == '\\') {
+                        throw SqlException.parserErr(len - 1, likeSeq, "LIKE pattern must not end with escape character");
+                    }
+
+                    int anyCount = countChar(likeSeq, '%');
+                    if (anyCount == 1) {
+                        if (len == 1) {
+                            return BooleanConstant.TRUE; // LIKE '%' case
+                        } else if (likeSeq.charAt(0) == '%') {
+                            // LIKE '%abc' case
+                            return new ConstEndsWithStrFunction(
+                                    value,
+                                    likeSeq.subSequence(1, len).toString(),
+                                    isCaseInsensitive()
+                            );
+                        } else if (likeSeq.charAt(len - 1) == '%') {
+                            // LIKE 'abc%' case
+                            return new ConstStartsWithStrFunction(
+                                    value,
+                                    likeSeq.subSequence(0, len - 1).toString(),
+                                    isCaseInsensitive()
+                            );
+                        }
+                    } else if (anyCount == 2) {
+                        if (len == 2) {
+                            return BooleanConstant.TRUE; // LIKE '%%' case
+                        } else if (likeSeq.charAt(0) == '%' && likeSeq.charAt(len - 1) == '%') {
+                            // LIKE '%abc%' case
+                            return new ConstContainsStrFunction(
+                                    value,
+                                    likeSeq.subSequence(1, len - 1).toString(),
+                                    isCaseInsensitive()
+                            );
+                        }
+                    }
+                }
+
+                String p = escapeSpecialChars(likeSeq, null);
                 assert p != null;
                 int flags = Pattern.DOTALL;
                 if (isCaseInsensitive()) {
@@ -115,6 +158,16 @@ public abstract class AbstractLikeStrFunctionFactory implements FunctionFactory 
         }
 
         throw SqlException.$(argPositions.getQuick(1), "use constant or bind variable");
+    }
+
+    private static int countChar(@NotNull CharSequence seq, char c) {
+        int count = 0;
+        for (int i = 0, n = seq.length(); i < n; i++) {
+            if (seq.charAt(i) == c) {
+                count++;
+            }
+        }
+        return count;
     }
 
     protected abstract boolean isCaseInsensitive();
@@ -176,12 +229,102 @@ public abstract class AbstractLikeStrFunctionFactory implements FunctionFactory 
         @Override
         public void toPlan(PlanSink sink) {
             sink.val(value);
-            //impl is regex 
+            // impl is regex
             sink.val(" ~ ");
             sink.val(pattern);
             if (!caseInsensitive) {
                 sink.val(" [case-sensitive]");
             }
+        }
+    }
+
+    private static class ConstContainsStrFunction extends BooleanFunction implements UnaryFunction {
+        private final boolean caseInsensitive;
+        private final String pattern;
+        private final Function value;
+
+        public ConstContainsStrFunction(Function value, String pattern, boolean caseInsensitive) {
+            this.value = value;
+            this.pattern = pattern;
+            this.caseInsensitive = caseInsensitive;
+        }
+
+        @Override
+        public Function getArg() {
+            return value;
+        }
+
+        @Override
+        public boolean getBool(Record rec) {
+            CharSequence cs = getArg().getStr(rec);
+            if (cs == null) {
+                return false;
+            }
+            if (caseInsensitive) {
+                return Chars.containsIgnoreCase(cs, pattern);
+            }
+            return Chars.contains(cs, pattern);
+        }
+
+        @Override
+        public boolean isReadThreadSafe() {
+            return true;
+        }
+
+        @Override
+        public void toPlan(PlanSink sink) {
+            sink.val(value);
+            if (caseInsensitive) {
+                sink.val(" ILIKE ");
+            } else {
+                sink.val(" LIKE ");
+            }
+            sink.val('%');
+            sink.val(pattern);
+            sink.val('%');
+        }
+    }
+
+    private static class ConstEndsWithStrFunction extends BooleanFunction implements UnaryFunction {
+        private final boolean caseInsensitive;
+        private final String pattern;
+        private final Function value;
+
+        public ConstEndsWithStrFunction(Function value, String pattern, boolean caseInsensitive) {
+            this.value = value;
+            this.pattern = pattern;
+            this.caseInsensitive = caseInsensitive;
+        }
+
+        @Override
+        public Function getArg() {
+            return value;
+        }
+
+        @Override
+        public boolean getBool(Record rec) {
+            CharSequence cs = getArg().getStr(rec);
+            if (caseInsensitive) {
+                return Chars.endsWithIgnoreCase(cs, pattern);
+            }
+            return Chars.endsWith(cs, pattern);
+        }
+
+        @Override
+        public boolean isReadThreadSafe() {
+            return true;
+        }
+
+        @Override
+        public void toPlan(PlanSink sink) {
+            sink.val(value);
+            if (caseInsensitive) {
+                sink.val(" ILIKE ");
+            } else {
+                sink.val(" LIKE ");
+            }
+            sink.val('%');
+            sink.val(pattern);
         }
     }
 
@@ -213,12 +356,55 @@ public abstract class AbstractLikeStrFunctionFactory implements FunctionFactory 
         @Override
         public void toPlan(PlanSink sink) {
             sink.val(value);
-            //impl is regex 
+            // impl is regex
             sink.val(" ~ ");
             sink.val(matcher.pattern().toString());
             if ((matcher.pattern().flags() & Pattern.CASE_INSENSITIVE) != 0) {
                 sink.val(" [case-sensitive]");
             }
+        }
+    }
+
+    private static class ConstStartsWithStrFunction extends BooleanFunction implements UnaryFunction {
+        private final boolean caseInsensitive;
+        private final String pattern;
+        private final Function value;
+
+        public ConstStartsWithStrFunction(Function value, String pattern, boolean caseInsensitive) {
+            this.value = value;
+            this.pattern = pattern;
+            this.caseInsensitive = caseInsensitive;
+        }
+
+        @Override
+        public Function getArg() {
+            return value;
+        }
+
+        @Override
+        public boolean getBool(Record rec) {
+            CharSequence cs = getArg().getStr(rec);
+            if (caseInsensitive) {
+                return Chars.startsWithIgnoreCase(cs, pattern);
+            }
+            return Chars.startsWith(cs, pattern);
+        }
+
+        @Override
+        public boolean isReadThreadSafe() {
+            return true;
+        }
+
+        @Override
+        public void toPlan(PlanSink sink) {
+            sink.val(value);
+            if (caseInsensitive) {
+                sink.val(" ILIKE ");
+            } else {
+                sink.val(" LIKE ");
+            }
+            sink.val(pattern);
+            sink.val('%');
         }
     }
 }