… core/src/main/java/io/questdb/std/str/Utf8s.java
… core/src/main/resources/io/questdb/bin/freebsd/libquestdb.so
… core/src/main/resources/io/questdb/bin/linux/libquestdb.so
… core/src/test/java/io/questdb/test/griffin/ExplainPlanTest.java
… core/src/test/java/io/questdb/test/griffin/engine/functions/regex/ILikeFunctionFactoryTest.java
… core/src/test/java/io/questdb/test/griffin/engine/functions/regex/LikeFunctionFactoryTest.java
… core/src/test/java/io/questdb/test/std/str/Utf8sTest.java