Hot runs for ClickBench queries:
```sql
-- Before: 15s
-- After: 3s
SELECT COUNT(*) FROM hits WHERE URL LIKE '%google%';

-- Before: 18s
-- After: 3s
SELECT SearchPhrase, MIN(URL), COUNT(*) AS c FROM hits WHERE URL LIKE '%google%' AND SearchPhrase IS NOT NULL GROUP BY SearchPhrase ORDER BY c DESC LIMIT 10;

-- Before: 17s
-- After: 2s
SELECT SearchPhrase, MIN(URL), MIN(Title), COUNT(*) AS c, count_distinct(UserID) FROM hits WHERE Title LIKE '%Google%' AND URL NOT LIKE '%.google.%' AND SearchPhrase IS NOT NULL GROUP BY SearchPhrase ORDER BY c DESC LIMIT 10;

-- Before: 5s
-- After: 1s
SELECT * FROM hits WHERE URL LIKE '%google%' ORDER BY EventTime LIMIT 10;
```