@ideoma 2023-10-27
#### [PR Coverage check]

:heart_eyes: **pass** : 115 / 159 (72.33%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/Chars.java|45|66|68.18%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/regex/AbstractLikeStrFunctionFactory.java|70|93|75.27%|