@ideoma 2023-10-31
#### [PR Coverage check]

:heart_eyes: **pass** : 167 / 189 (88.36%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/Chars.java|45|66|68.18%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/regex/AbstractLikeStrFunctionFactory.java|122|123|99.19%|