@ideoma 2023-10-31
#### [PR Coverage check]

:heart_eyes: **pass** : 171 / 193 (88.60%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/Chars.java|49|70|70.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/regex/AbstractLikeStrFunctionFactory.java|122|123|99.19%|