@puzpuzpuz 2023-11-01
@bluestreak01 I've pushed a couple of commits with alternative `indexOf()` implementations, as well as a simple microbenchmark.

The first commit uses Two-way string-matching algorithm (https://en.wikipedia.org/wiki/Two-way_string-matching_algorithm):
```
Benchmark                                       (len)  Mode  Cnt     Score     Error  Units
CharsIndexOfBenchmark.testNaiveIndexOf             32  avgt    3    21.461 ±  15.441  ns/op
CharsIndexOfBenchmark.testNaiveIndexOf            128  avgt    3    63.101 ±   1.812  ns/op
CharsIndexOfBenchmark.testNaiveIndexOf           1024  avgt    3   412.373 ± 209.485  ns/op
CharsIndexOfBenchmark.testStdlibIndexOf            32  avgt    3     5.497 ±   0.504  ns/op
CharsIndexOfBenchmark.testStdlibIndexOf           128  avgt    3    13.705 ±   0.704  ns/op
CharsIndexOfBenchmark.testStdlibIndexOf          1024  avgt    3    90.498 ±   9.719  ns/op
CharsIndexOfBenchmark.testTwoWayStringMatching     32  avgt    3    79.904 ±  15.639  ns/op
CharsIndexOfBenchmark.testTwoWayStringMatching    128  avgt    3   224.817 ±  19.377  ns/op
CharsIndexOfBenchmark.testTwoWayStringMatching   1024  avgt    3  1573.806 ±  66.869  ns/op
```

Here `testNaiveIndexOf` stands for the older version of `indexOf()` and `testStdlibIndexOf` stands for `String#indexOf()`.

The next one is based on `StringUTF16#indexOfUnsafe()` (open JDK Java impl of `indexOf()`):
```
Benchmark                                       (len)  Mode  Cnt    Score   Error  Units
CharsIndexOfBenchmark.testBidirectionalIndexOf     32  avgt    3   20.930 ± 1.540  ns/op
CharsIndexOfBenchmark.testBidirectionalIndexOf    128  avgt    3   63.764 ± 2.308  ns/op
CharsIndexOfBenchmark.testBidirectionalIndexOf   1024  avgt    3  404.783 ± 9.569  ns/op
CharsIndexOfBenchmark.testIndexOf                  32  avgt    3   13.331 ± 0.448  ns/op
CharsIndexOfBenchmark.testIndexOf                 128  avgt    3   25.695 ± 0.669  ns/op
CharsIndexOfBenchmark.testIndexOf                1024  avgt    3  147.501 ± 7.782  ns/op
CharsIndexOfBenchmark.testStdlibIndexOf            32  avgt    3    5.514 ± 0.217  ns/op
CharsIndexOfBenchmark.testStdlibIndexOf           128  avgt    3   13.829 ± 1.199  ns/op
CharsIndexOfBenchmark.testStdlibIndexOf          1024  avgt    3   89.912 ± 2.709  ns/op
```

Here `testBidirectionalIndexOf` stands for the old impl and `testIndexOf` stands for the new one. I kept this one as it looks faster than the Two-way string-matching algo, at least in this very scenario.

I'm going to check if there is any improvements for the ClickBench dataset.