@ideoma 2023-11-01
#### [PR Coverage check]

:heart_eyes: **pass** : 223 / 231 (96.54%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/str/Utf8s.java|24|26|92.31%|
|:large_blue_circle:|io/questdb/std/Chars.java|77|82|93.90%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/regex/AbstractLikeStrFunctionFactory.java|122|123|99.19%|