@puzpuzpuz 2023-11-01
Updated the query exec times in the PR description. The updated `indexOf` improved them, at least on my machine.